Refactor Homes from index.js into game.js
