+++ lib/game.js
@@ -15,7 +15,7 @@ class Game {
     this.slowCars = [];
     this.fastCars = [];
     this.semis = [];
-    this.homes = [];
+    this.homes = this.buildHomes();
     this.toad = new Toad(475, 575, 'safe', this.homes)
   }
 
@@ -49,6 +49,9 @@ class Game {
     }, this)
     this.logs.forEach(function(platform) {
       platform.draw(this.canvas, this.context);
+    }, this)
+    this.homes.forEach(function(home) {
+      home.draw(this.canvas, this.context);
     }, this)
   }
 
@@ -207,6 +210,18 @@ class Game {
     return buildRiver5Array;
   }
 
+  buildHomes() {
+    let homeYcoord = 25;
+    let startXcoord = this.squareWidth - 2.5;
+    let gap = this.squareWidth * 2;
+    let numberOfHomes = 5;
+    let buildHomeArray = [];
+
+    for (let i = 0; i < numberOfHomes; i++) {
+      buildHomeArray.push(new Home(startXcoord + (this.squareWidth + gap) * i, homeYcoord, this.squareWidth, this.squareHeight));
+    }
+    return buildHomeArray;
+  }
 }
 
 module.exports = Game;
