+++ lib/index.js
@@ -1,49 +1,21 @@
 let style = require('./style.css');
-const Toad = require('./toad.js');
-const Home = require('./home.js');
-const {Platform, Turtle, Crocodile} = require('./platform.js');
-const {Obstacle, Auto, Tractor, SlowCar, Semi, FastCar, Snake} = require('./obstacle.js');
 let Game = require('./game.js');
-
-var canvas = document.getElementById('canvas');
-var context = canvas.getContext('2d');
-
+let canvas = document.getElementById('canvas');
+let context = canvas.getContext('2d');
 let game = new Game(canvas, context);
 
-let squareHeight = 50;
-let squareWidth = 67;
-let homes = buildHomes();
-
+document.addEventListener('keyup', evalInput);
 game.buildLevel();
-// let newToad = new Toad(475, 575, 'safe', homes);
 
 requestAnimationFrame(function gameLoop() {
   game.drawBackground();
   game.drawGamePieces();
-  homes.forEach(function(home) {
-    home.draw(canvas, context);
-  })
   collisionDetection();
   game.toad.drawToad(context);
   requestAnimationFrame(gameLoop);
 })
 
-document.addEventListener('keyup', evalInput);
-
 // functions
-
-function buildHomes() {
-  let homeYcoord = 25;
-  let startXcoord = squareWidth - 2.5;
-  let gap = squareWidth * 2;
-  let numberOfHomes = 5;
-  let buildHomeArray = [];
-
-  for (let i = 0; i < numberOfHomes; i++) {
-    buildHomeArray.push(new Home(startXcoord + (squareWidth + gap) * i, homeYcoord, squareWidth, squareHeight));
-  }
-  return buildHomeArray;
-}
 
 function collisionDetection() {
   let testArray = game.tractors.concat(game.slowCars).concat(game.fastCars).concat(game.semis);
